+++ internal/http/middleware/util.go
@@ -1,0 +1,12 @@
+package middleware
+
+import "strings"
+
+func stringContains(slice []string, val string) bool {
+	for _, item := range slice {
+		if strings.Contains(val, item) {
+			return true
+		}
+	}
+	return false
+}